requirements.txt
openai
openai-agents
python-dotenv
jsonschema
redis
a2a-sdk
sqlalchemy
azure-ai-projects
azure-identity
azure-ai-agents